// src/app/models/admin/ThemeSetting/ThemeSettingDTO.ts
export class ThemeSettingDTO {
    Id: number = 0;
    Logo: string = '';
    Favicon: string = '';
    MainBackgorundColor: string = '';
    MainTextColor: string = '';
    MainLabel: string = '';
    TitleAr: string = '';
    TitleEn: string = '';
    LogoFileContent: string = '';
    LogoFileExtension: string = '';
    FaviconFileContent: string = '';
  }